+++ air/src/contexts/execution_trace_context.rs
@@ -17,6 +17,7 @@
 mod executed_state;
 
 pub use executed_state::CallResult;
+pub use executed_state::ParResult;
 pub use executed_state::ExecutedState;
 
 use serde::Deserialize;
