+++ air/src/contexts/execution_trace_context/executed_state.rs
@@ -20,6 +20,10 @@ use serde::Deserialize;
 use serde::Serialize;
 use std::rc::Rc;
 
+#[derive(Debug, Default, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub struct ParResult(pub usize, pub usize);
+
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 pub enum CallResult {
@@ -36,8 +40,14 @@ pub enum CallResult {
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 pub enum ExecutedState {
-    Par(usize, usize),
+    Par(ParResult),
     Call(CallResult),
+}
+
+impl ExecutedState {
+    pub fn par(left: usize, right: usize) -> Self {
+        Self::Par(ParResult(left, right))
+    }
 }
 
 impl std::fmt::Display for ExecutedState {
@@ -46,7 +56,7 @@ impl std::fmt::Display for ExecutedState {
         use ExecutedState::*;
 
         match self {
-            Par(left, right) => write!(f, "Par({}, {})", left, right),
+            Par(ParResult(left, right)) => write!(f, "Par({}, {})", left, right),
             Call(RequestSentBy(peer_id)) => write!(f, "RequestSentBy({})", peer_id),
             Call(Executed(result)) => write!(f, "Executed({:?})", result),
             Call(CallServiceFailed(ret_code, err_msg)) => write!(f, "CallServiceFailed({}, {})", ret_code, err_msg),
