+++ air/src/execution/air/call.rs
@@ -65,178 +65,3 @@ fn set_last_error<'i>(
     exec_ctx.last_error = Some(last_error);
     exec_ctx.last_error_could_be_set = false;
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::contexts::execution_trace::ExecutionTrace;
-    use crate::JValue;
-
-    use air_test_utils::call_vm;
-    use air_test_utils::create_avm;
-    use air_test_utils::echo_string_call_service;
-    use air_test_utils::set_variable_call_service;
-    use air_test_utils::unit_call_service;
-    use air_test_utils::CallServiceClosure;
-    use air_test_utils::IValue;
-    use air_test_utils::NEVec;
-
-    use std::rc::Rc;
-
-    // Check that %init_peer_id% alias works correctly (by comparing result with it and explicit peer id).
-    // Additionally, check that empty string for data does the same as empty call path.
-    #[test]
-    fn current_peer_id_call() {
-        use crate::contexts::execution_trace::CallResult::*;
-        use crate::contexts::execution_trace::ExecutedState::*;
-
-        let vm_peer_id = String::from("test_peer_id");
-        let mut vm = create_avm(unit_call_service(), vm_peer_id.clone());
-
-        let service_id = String::from("local_service_id");
-        let function_name = String::from("local_fn_name");
-        let script = format!(
-            r#"
-               (call %init_peer_id% ("{}" "{}") [] result_name)
-            "#,
-            service_id, function_name
-        );
-
-        let res = call_vm!(vm, vm_peer_id.clone(), script.clone(), "[]", "[]");
-        let call_path: ExecutionTrace = serde_json::from_slice(&res.data).expect("should be a valid json");
-
-        let executed_call_state = Call(Executed(Rc::new(JValue::String(String::from("test")))));
-        assert_eq!(call_path.len(), 1);
-        assert_eq!(call_path[0], executed_call_state);
-        assert!(res.next_peer_pks.is_empty());
-
-        let script = format!(
-            r#"
-               (call "{}" ("{}" "{}") [] result_name)
-            "#,
-            vm_peer_id, service_id, function_name
-        );
-
-        let res = call_vm!(vm, "asd", script.clone(), "[]", "[]");
-
-        // test that empty string for data works
-        let res_with_empty_string = call_vm!(vm, "asd", script, "", "");
-        assert_eq!(res_with_empty_string, res);
-    }
-
-    // Check that specifying remote peer id in call will result its appearing in next_peer_pks.
-    #[test]
-    fn remote_peer_id_call() {
-        use crate::contexts::execution_trace::CallResult::*;
-        use crate::contexts::execution_trace::ExecutedState::*;
-
-        let some_local_peer_id = String::from("some_local_peer_id");
-        let mut vm = create_avm(echo_string_call_service(), some_local_peer_id.clone());
-
-        let remote_peer_id = String::from("some_remote_peer_id");
-        let script = format!(
-            r#"(call "{}" ("local_service_id" "local_fn_name") ["arg"] result_name)"#,
-            remote_peer_id
-        );
-
-        let res = call_vm!(vm, "asd", script, "[]", "[]");
-        let call_path: ExecutionTrace = serde_json::from_slice(&res.data).expect("should be a valid json");
-
-        assert_eq!(call_path.len(), 1);
-        assert_eq!(call_path[0], Call(RequestSentBy(some_local_peer_id)));
-        assert_eq!(res.next_peer_pks, vec![remote_peer_id]);
-    }
-
-    // Check that setting variables works as expected.
-    #[test]
-    fn variables() {
-        let mut vm = create_avm(unit_call_service(), "remote_peer_id");
-        let mut set_variable_vm = create_avm(set_variable_call_service(r#""remote_peer_id""#), "set_variable");
-
-        let script = format!(
-            r#"
-            (seq 
-                (call "set_variable" ("some_service_id" "local_fn_name") [] remote_peer_id)
-                (call remote_peer_id ("some_service_id" "local_fn_name") [] result_name)
-            )
-        "#,
-        );
-
-        let res = call_vm!(set_variable_vm, "asd", script.clone(), "[]", "[]");
-        let res = call_vm!(vm, "asd", script, "[]", res.data);
-
-        assert!(res.next_peer_pks.is_empty());
-    }
-
-    // Check that duplicate variables are impossible.
-    #[test]
-    fn duplicate_variables() {
-        let mut vm = create_avm(unit_call_service(), "some_peer_id");
-
-        let script = format!(
-            r#"
-            (seq
-                (call "some_peer_id" ("some_service_id" "local_fn_name") [] modules)
-                (call "some_peer_id" ("some_service_id" "local_fn_name") [] modules)
-            )
-        "#,
-        );
-
-        let res = call_vm!(vm, "asd", script, "", "");
-
-        assert_eq!(res.ret_code, 1005);
-        assert!(res.next_peer_pks.is_empty());
-    }
-
-    // Check that string literals can be used as call parameters.
-    #[test]
-    fn string_parameters() {
-        use crate::contexts::execution_trace::CallResult::*;
-        use crate::contexts::execution_trace::ExecutedState::*;
-
-        let call_service: CallServiceClosure = Box::new(|_, args| -> Option<IValue> {
-            let arg = match &args[2] {
-                IValue::String(str) => str,
-                _ => unreachable!(),
-            };
-
-            Some(IValue::Record(
-                NEVec::new(vec![IValue::S32(0), IValue::String(arg.clone())]).unwrap(),
-            ))
-        });
-
-        let vm_peer_id = String::from("A");
-        let mut vm = create_avm(call_service, vm_peer_id.clone());
-
-        let set_variable_vm_peer_id = String::from("set_variable");
-        let mut set_variable_vm = create_avm(
-            set_variable_call_service(r#""arg3_value""#),
-            set_variable_vm_peer_id.clone(),
-        );
-
-        let service_id = String::from("some_service_id");
-        let function_name = String::from("local_fn_name");
-        let script = format!(
-            r#"
-            (seq 
-                (call "{}" ("{}" "{}") [] arg3)
-                (call "{}" ("{}" "{}") ["arg1" "arg2" arg3] result)
-            )
-        "#,
-            set_variable_vm_peer_id, service_id, function_name, vm_peer_id, service_id, function_name
-        );
-
-        let res = call_vm!(set_variable_vm, "asd", script.clone(), "[]", "[]");
-        let res = call_vm!(vm, "asd", script, "[]", res.data);
-        let call_path: ExecutionTrace = serde_json::from_slice(&res.data).expect("should be a valid json");
-
-        assert_eq!(call_path.len(), 2);
-        assert_eq!(
-            call_path[1],
-            Call(Executed(Rc::new(JValue::Array(vec![
-                JValue::String(String::from("arg1")),
-                JValue::String(String::from("arg2")),
-                JValue::String(String::from("arg3_value")),
-            ]))))
-        );
-    }
-}
