+++ air/src/execution/air/par.rs
@@ -20,6 +20,7 @@ use super::ExecutionResult;
 use super::ExecutionTraceCtx;
 use super::Instruction;
 use crate::contexts::execution_trace::ExecutedState;
+use crate::contexts::execution_trace::ParResult;
 use crate::log_instruction;
 use crate::log_targets::EXECUTED_STATE_CHANGING;
 
@@ -48,7 +49,7 @@ impl<'i> ExecutableInstruction<'i> for Par<'i> {
         let (left_subtree_size, right_subtree_size) = extract_subtree_sizes(trace_ctx)?;
 
         let par_pos = trace_ctx.new_trace.len();
-        trace_ctx.new_trace.push_back(ExecutedState::Par(0, 0));
+        trace_ctx.new_trace.push_back(ExecutedState::par(0, 0));
 
         // execute a left subtree of this par
         execute_subtree(&self.0, left_subtree_size, exec_ctx, trace_ctx, par_pos, Left)?;
@@ -82,7 +83,7 @@ fn extract_subtree_sizes(trace_ctx: &mut ExecutionTraceCtx) -> ExecutionResult<(
 
     // unwrap is safe here because of length's been checked
     match trace_ctx.current_trace.pop_front().unwrap() {
-        ExecutedState::Par(left, right) => Ok((left, right)),
+        ExecutedState::Par(ParResult(left, right)) => Ok((left, right)),
         state => crate::exec_err!(InvalidExecutedState(String::from("par"), state)),
     }
 }
@@ -146,7 +147,7 @@ fn update_par_state(
     // unwrap is safe here, because this par is added at the beginning of this par instruction.
     let par_state = trace_ctx.new_trace.get_mut(current_par_pos).unwrap();
     match par_state {
-        ExecutedState::Par(left, right) => {
+        ExecutedState::Par(ParResult(left, right)) => {
             if let SubtreeType::Left = subtree_type {
                 *left = new_subtree_size;
             } else {
@@ -161,54 +162,5 @@ fn update_par_state(
             );
         }
         _ => unreachable!("current_pas_pos must point to a par state"),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use air_test_utils::call_vm;
-    use air_test_utils::create_avm;
-    use air_test_utils::unit_call_service;
-
-    #[test]
-    fn par_remote_remote() {
-        use std::collections::HashSet;
-
-        let mut vm = create_avm(unit_call_service(), "");
-
-        let script = String::from(
-            r#"
-            (par 
-                (call "remote_peer_id_1" ("local_service_id" "local_fn_name") [] result_name)
-                (call "remote_peer_id_2" ("service_id" "fn_name") [] g)
-            )"#,
-        );
-
-        let mut res = call_vm!(vm, "", script, "[]", "[]");
-
-        let peers_result: HashSet<_> = res.next_peer_pks.drain(..).collect();
-        let peers_right: HashSet<_> =
-            maplit::hashset!(String::from("remote_peer_id_1"), String::from("remote_peer_id_2"));
-
-        assert_eq!(peers_result, peers_right);
-    }
-
-    #[test]
-    fn par_local_remote() {
-        let local_peer_id = "local_peer_id";
-        let mut vm = create_avm(unit_call_service(), local_peer_id);
-
-        let script = format!(
-            r#"
-            (par 
-                (call "{}" ("local_service_id" "local_fn_name") [] result_name)
-                (call "remote_peer_id_2" ("service_id" "fn_name") [] g)
-            )"#,
-            local_peer_id
-        );
-
-        let res = call_vm!(vm, "", script, "[]", "[]");
-
-        assert_eq!(res.next_peer_pks, vec![String::from("remote_peer_id_2")]);
     }
 }
