+++ air/src/execution/air/xor.rs
@@ -56,206 +56,3 @@ fn print_xor_log(e: &ExecutionError) {
         e => log::warn!("xor caught an error: {}", e),
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::contexts::execution_trace::ExecutionTrace;
-    use crate::JValue;
-
-    use air_test_utils::call_vm;
-    use air_test_utils::create_avm;
-    use air_test_utils::fallible_call_service;
-
-    use std::rc::Rc;
-
-    #[test]
-    fn xor() {
-        use crate::contexts::execution_trace::CallResult::*;
-        use crate::contexts::execution_trace::ExecutedState::*;
-
-        let local_peer_id = "local_peer_id";
-        let fallible_service_id = String::from("service_id_1");
-        let mut vm = create_avm(fallible_call_service(fallible_service_id), local_peer_id);
-
-        let script = format!(
-            r#"
-            (xor
-                (call "{0}" ("service_id_1" "local_fn_name") [] result_1)
-                (call "{0}" ("service_id_2" "local_fn_name") [] result_2)
-            )"#,
-            local_peer_id,
-        );
-
-        let res = call_vm!(vm, "asd", script, "[]", "[]");
-        let actual_trace: ExecutionTrace = serde_json::from_slice(&res.data).expect("should be valid json");
-        let executed_call_result = Call(Executed(Rc::new(JValue::String(String::from("res")))));
-
-        assert_eq!(actual_trace.len(), 2);
-        assert_eq!(
-            actual_trace[0],
-            Call(CallServiceFailed(1, Rc::new(String::from(r#"error"#))))
-        );
-        assert_eq!(actual_trace[1], executed_call_result);
-
-        let script = format!(
-            r#"
-            (xor
-                (call "{0}" ("service_id_2" "local_fn_name") [] result_1)
-                (call "{0}" ("service_id_1" "local_fn_name") [] result_2)
-            )"#,
-            local_peer_id
-        );
-
-        let res = call_vm!(vm, "asd", script, "[]", "[]");
-        let actual_trace: ExecutionTrace = serde_json::from_slice(&res.data).expect("should be valid json");
-
-        assert_eq!(actual_trace.len(), 1);
-        assert_eq!(actual_trace[0], executed_call_result);
-    }
-
-    #[test]
-    fn xor_var_not_found() {
-        use crate::contexts::execution_trace::CallResult::*;
-        use crate::contexts::execution_trace::ExecutedState::*;
-        use air_test_utils::echo_string_call_service;
-
-        let local_peer_id = "local_peer_id";
-        let mut vm = create_avm(echo_string_call_service(), local_peer_id);
-
-        let script = format!(
-            r#"
-            (xor
-                (par
-                    (call "unknown_peer" ("service_id_1" "local_fn_name") [] lazy_defined_variable)
-                    (call "{0}" ("service_id_1" "local_fn_name") [lazy_defined_variable] result)
-                )
-                (call "{0}" ("service_id_2" "local_fn_name") ["expected"] result)
-            )"#,
-            local_peer_id,
-        );
-
-        let res = call_vm!(vm, "asd", script, "[]", "[]");
-        let actual_trace: ExecutionTrace = serde_json::from_slice(&res.data).expect("should be valid json");
-        assert_eq!(actual_trace[0], Par(1, 0));
-        assert_eq!(actual_trace[1], Call(RequestSentBy(String::from("local_peer_id"))));
-    }
-
-    #[test]
-    fn xor_multiple_variables_found() {
-        use crate::contexts::execution_trace::CallResult::*;
-        use crate::contexts::execution_trace::ExecutedState::*;
-        use air_test_utils::echo_string_call_service;
-
-        let set_variables_peer_id = "set_variables_peer_id";
-        let mut set_variables_vm = create_avm(echo_string_call_service(), set_variables_peer_id);
-
-        let local_peer_id = "local_peer_id";
-        let mut vm = create_avm(echo_string_call_service(), local_peer_id);
-
-        let test_string_1 = String::from("some_string");
-        let test_string_2 = String::from("expected_string");
-        let script = format!(
-            r#"
-            (seq
-                (call "{0}" ("service_id_1" "local_fn_name") ["{2}"] result_1)
-                (xor
-                    (call "{1}" ("service_id_1" "local_fn_name") [""] result_1)
-                    (call "{1}" ("service_id_2" "local_fn_name") ["{3}"] result_2)
-                )
-            )"#,
-            set_variables_peer_id, local_peer_id, test_string_1, test_string_2
-        );
-
-        let res = call_vm!(set_variables_vm, "asd", script.clone(), "[]", "[]");
-        let res = call_vm!(vm, "asd", script, "[]", res.data);
-        let actual_trace: ExecutionTrace = serde_json::from_slice(&res.data).expect("should be valid json");
-
-        assert_eq!(actual_trace.len(), 2);
-        assert_eq!(actual_trace[0], Call(Executed(Rc::new(JValue::String(test_string_1)))));
-        assert_eq!(actual_trace[1], Call(Executed(Rc::new(JValue::String(test_string_2)))));
-    }
-
-    #[test]
-    fn xor_par() {
-        use crate::contexts::execution_trace::CallResult::*;
-        use crate::contexts::execution_trace::ExecutedState::*;
-
-        let fallible_service_id = String::from("service_id_1");
-        let local_peer_id = "local_peer_id";
-        let mut vm = create_avm(fallible_call_service(fallible_service_id.clone()), local_peer_id);
-
-        let script = format!(
-            r#"
-            (xor
-                (par
-                    (seq
-                        (call "{0}" ("service_id_2" "local_fn_name") [] result_1)
-                        (call "{0}" ("service_id_2" "local_fn_name") [] result_2)
-                    )
-                    (par
-                        (call "{0}" ("service_id_1" "local_fn_name") [] result_3)
-                        (call "{0}" ("service_id_2" "local_fn_name") [] result_4)
-                    )
-                )
-                (seq
-                    (call "{0}" ("service_id_2" "local_fn_name") [] result_4)
-                    (call "{0}" ("service_id_2" "local_fn_name") [] result_5)
-                )
-            )"#,
-            local_peer_id
-        );
-
-        let result = call_vm!(vm, "asd", script.clone(), "[]", "[]");
-        let actual_trace: ExecutionTrace = serde_json::from_slice(&result.data).expect("should be valid json");
-
-        let res = String::from("res");
-        let executed_call_result = Rc::new(JValue::String(res));
-
-        let expected_trace = vec![
-            Par(2, 2),
-            Call(Executed(executed_call_result.clone())),
-            Call(Executed(executed_call_result.clone())),
-            Par(1, 0),
-            Call(CallServiceFailed(1, Rc::new(String::from(r#"error"#)))),
-            Call(Executed(executed_call_result.clone())),
-            Call(Executed(executed_call_result.clone())),
-        ];
-
-        assert_eq!(actual_trace, expected_trace);
-
-        let result = call_vm!(vm, "asd", script, "[]", result.data);
-        let actual_trace: ExecutionTrace = serde_json::from_slice(&result.data).expect("should be valid json");
-        assert_eq!(actual_trace, expected_trace);
-    }
-
-    #[test]
-    fn last_error_with_xor() {
-        use crate::contexts::execution_trace::CallResult::*;
-        use crate::contexts::execution_trace::ExecutedState::*;
-        use air_test_utils::echo_string_call_service;
-
-        let faillible_peer_id = "failible_peer_id";
-        let mut faillible_vm = create_avm(fallible_call_service("service_id_1"), faillible_peer_id);
-        let local_peer_id = "local_peer_id";
-        let mut vm = create_avm(echo_string_call_service(), local_peer_id);
-
-        let script = format!(
-            r#"
-            (xor
-                (call "{0}" ("service_id_1" "local_fn_name") [] result)
-                (call "{1}" ("service_id_2" "local_fn_name") [%last_error%] result)
-            )"#,
-            faillible_peer_id, local_peer_id,
-        );
-
-        let res = call_vm!(faillible_vm, "asd", script.clone(), "", "");
-        let res = call_vm!(vm, "asd", script, "", res.data);
-        let actual_trace: ExecutionTrace = serde_json::from_slice(&res.data).expect("should be valid json");
-
-        let expected_state = Call(Executed(Rc::new(JValue::String(String::from(
-            "{\"error\":\"Local service error: ret_code is 1, error message is \'error\'\",\"instruction\":\"call \\\"failible_peer_id\\\" (\\\"service_id_1\\\" \\\"local_fn_name\\\") [] result\"}"
-        )))));
-
-        assert_eq!(actual_trace[1], expected_state);
-    }
-}
