+++ air/src/preparation/mod.rs
@@ -17,6 +17,8 @@
 mod data_merging;
 mod errors;
 mod preparation;
+#[cfg(tests)]
+mod tests;
 
 pub(crate) use errors::DataMergingError;
 pub(crate) use errors::PreparationError;
@@ -25,6 +27,7 @@ pub(crate) use preparation::PreparationDescriptor;
 
 pub(self) use crate::contexts::execution::*;
 pub(self) use crate::contexts::execution_trace::CallResult;
+pub(self) use crate::contexts::execution_trace::ParResult;
 pub(self) use crate::contexts::execution_trace::ExecutedState;
 pub(self) use crate::contexts::execution_trace::ExecutionTrace;
 pub(self) use crate::contexts::execution_trace::ExecutionTraceCtx;
