--- air/tests/build_test_binaries.sh
@@ -1,10 +1,0 @@
-#!/usr/bin/env bash
-
-set -euo pipefail
-
-for dir in ./security_tetraplets/*; do
-    # skip non-directory entries
-    [ -d "$dir" ] || continue
-
-    (cd "$dir"; marine build)
-done
