--- air/tests/security_tetraplets/log_storage/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "log_storage"
-version = "0.1.0"
-authors = ["Fluence Labs"]
-edition = "2018"
-
-[[bin]]
-name = "log_storage"
-path = "src/main.rs"
-
-[dependencies]
-air = { path = "../../../../air" }
-fluence = "0.6.5"
-
-[workspace]
