+++ air/tests/test_module/integration/join.rs
@@ -14,17 +14,16 @@
  * limitations under the License.
  */
 
+use air::execution_trace::ExecutionTrace;
 use air_test_utils::call_vm;
 use air_test_utils::create_avm;
+use air_test_utils::executed_state;
 use air_test_utils::unit_call_service;
 use air_test_utils::CallServiceClosure;
 use air_test_utils::IValue;
 use air_test_utils::NEVec;
 
-use pretty_assertions::assert_eq;
 use serde_json::json;
-
-type JValue = serde_json::Value;
 
 #[test]
 fn join_chat() {
@@ -55,7 +54,7 @@ fn join_chat() {
                     (seq
                         (call "Remote" ("920e3ba3-cbdf-4ae3-8972-0fa2f31fffd9" "get_users") [] members)
                         (fold members m
-                            (par 
+                            (par
                                 (seq
                                     (call m.$.[1]! ("identity" "") [] $void)
                                     (call m.$.[0]! ("fgemb3" "add") [] $void3)
@@ -71,43 +70,41 @@ fn join_chat() {
 
     let client_1_res = call_vm!(client_1, "asd", script.clone(), "[]", "[]");
 
-    let client_1_res_json: JValue =
+    let client_1_actual_trace: ExecutionTrace =
         serde_json::from_slice(&client_1_res.data).expect("interpreter should return valid json");
 
-    let client_1_expected_json = json!([
-        { "call": {"request_sent_by": "A" } },
-    ]);
+    let client_1_expected_trace = vec![executed_state::request_sent_by("A")];
 
-    assert_eq!(client_1_res_json, client_1_expected_json);
+    assert_eq!(client_1_actual_trace, client_1_expected_trace);
     assert_eq!(client_1_res.next_peer_pks, vec![String::from("Relay1")]);
 
     let relay_1_res = call_vm!(relay_1, "asd", script.clone(), client_1_res.data, "[]");
 
-    let relay_1_res_json: JValue =
+    let relay_1_actual_trace: ExecutionTrace =
         serde_json::from_slice(&relay_1_res.data).expect("interpreter should return valid json");
 
-    let relay_1_expected_json = json!( [
-        { "call": { "executed" : "test" } },
-        { "call": { "request_sent_by": "Relay1" } },
-    ]);
+    let relay_1_expected_trace = vec![
+        executed_state::stream_string("test", "void1"),
+        executed_state::request_sent_by("Relay1"),
+    ];
 
-    assert_eq!(relay_1_res_json, relay_1_expected_json);
+    assert_eq!(relay_1_actual_trace, relay_1_expected_trace);
     assert_eq!(relay_1_res.next_peer_pks, vec![String::from("Remote")]);
 
     let remote_res = call_vm!(remote, "asd", script.clone(), relay_1_res.data, "[]");
 
-    let remote_res_json: JValue =
+    let remote_actual_trace: ExecutionTrace =
         serde_json::from_slice(&remote_res.data).expect("interpreter should return valid json");
 
-    let remote_expected_json = json!( [
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : [["A", "Relay1"], ["B", "Relay2"]]} },
-        { "call": { "executed" : [["A", "Relay1"], ["B", "Relay2"]]} },
-        { "par": [1, 2] },
-        { "call": { "request_sent_by" : "Remote" } },
-        { "par": [1, 0] },
-        { "call": { "request_sent_by" : "Remote" } },
-    ]);
+    let remote_expected_trace = vec![
+        executed_state::stream_string("test", "void1"),
+        executed_state::stream_jvalue(json!([["A", "Relay1"], ["B", "Relay2"]]), "void2"),
+        executed_state::scalar_jvalue(json!([["A", "Relay1"], ["B", "Relay2"]])),
+        executed_state::par(1, 2),
+        executed_state::request_sent_by("Remote"),
+        executed_state::par(1, 0),
+        executed_state::request_sent_by("Remote"),
+    ];
 
     let remote_res_next_peer_pks: HashSet<_> = remote_res.next_peer_pks.iter().map(|s| s.as_str()).collect();
     let next_peer_pks_right = maplit::hashset! {
@@ -115,83 +112,82 @@ fn join_chat() {
         "Relay2",
     };
 
-    assert_eq!(remote_res_json, remote_expected_json);
+    assert_eq!(remote_actual_trace, remote_expected_trace);
     assert_eq!(remote_res_next_peer_pks, next_peer_pks_right);
 
     let relay_1_res = call_vm!(relay_1, "asd", script.clone(), remote_res.data.clone(), "[]");
 
-    let relay_1_res_json: JValue =
+    let relay_1_actual_trace: ExecutionTrace =
         serde_json::from_slice(&relay_1_res.data).expect("interpreter should return valid json");
+    let relay_1_expected_trace = vec![
+        executed_state::stream_string("test", "void1"),
+        executed_state::stream_jvalue(json!([["A", "Relay1"], ["B", "Relay2"]]), "void2"),
+        executed_state::scalar_jvalue(json!([["A", "Relay1"], ["B", "Relay2"]])),
+        executed_state::par(2, 2),
+        executed_state::stream_string("test", "void"),
+        executed_state::request_sent_by("Relay1"),
+        executed_state::par(1, 0),
+        executed_state::request_sent_by("Remote"),
+    ];
 
-    let relay_1_expected_json = json!( [
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : [["A", "Relay1"], ["B", "Relay2"]]} },
-        { "call": { "executed" : [["A", "Relay1"], ["B", "Relay2"]]} },
-        { "par": [2, 2] },
-        { "call": { "executed" : "test" } },
-        { "call": { "request_sent_by" : "Relay1" } },
-        { "par": [1, 0] },
-        { "call": { "request_sent_by" : "Remote" } },
-    ]);
-
-    assert_eq!(relay_1_res_json, relay_1_expected_json);
+    assert_eq!(relay_1_actual_trace, relay_1_expected_trace);
     assert_eq!(relay_1_res.next_peer_pks, vec![String::from("A")]);
 
     let client_1_res = call_vm!(client_1, "asd", script.clone(), relay_1_res.data, "[]");
 
-    let client_1_res_json: JValue =
+    let client_1_actual_trace: ExecutionTrace =
         serde_json::from_slice(&client_1_res.data).expect("interpreter should return valid json");
 
-    let client_1_expected_json = json!( [
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : [["A", "Relay1"], ["B", "Relay2"]]} },
-        { "call": { "executed" : [["A", "Relay1"], ["B", "Relay2"]]} },
-        { "par": [2, 2] },
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : "test" } },
-        { "par": [1, 0] },
-        { "call": { "request_sent_by" : "Remote" } },
-    ]);
+    let client_1_expected_trace = vec![
+        executed_state::stream_string("test", "void1"),
+        executed_state::stream_jvalue(json!([["A", "Relay1"], ["B", "Relay2"]]), "void2"),
+        executed_state::scalar_jvalue(json!([["A", "Relay1"], ["B", "Relay2"]])),
+        executed_state::par(2, 2),
+        executed_state::stream_string("test", "void"),
+        executed_state::stream_string("test", "void3"),
+        executed_state::par(1, 0),
+        executed_state::request_sent_by("Remote"),
+    ];
 
-    assert_eq!(client_1_res_json, client_1_expected_json);
+    assert_eq!(client_1_actual_trace, client_1_expected_trace);
     assert_eq!(client_1_res.next_peer_pks, Vec::<String>::new());
 
     let relay_2_res = call_vm!(relay_2, "asd", script.clone(), remote_res.data, "[]");
 
-    let relay_2_res_json: JValue =
+    let relay_2_actual_trace: ExecutionTrace =
         serde_json::from_slice(&relay_2_res.data).expect("interpreter should return valid json");
 
-    let relay_2_expected_json = json!( [
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : [["A", "Relay1"], ["B", "Relay2"]]} },
-        { "call": { "executed" : [["A", "Relay1"], ["B", "Relay2"]]} },
-        { "par": [1, 3] },
-        { "call": { "request_sent_by" : "Remote" } },
-        { "par": [2, 0] },
-        { "call": { "executed" : "test" } },
-        { "call": { "request_sent_by" : "Relay2" } },
-    ]);
+    let relay_2_expected_trace = vec![
+        executed_state::stream_string("test", "void1"),
+        executed_state::stream_jvalue(json!([["A", "Relay1"], ["B", "Relay2"]]), "void2"),
+        executed_state::scalar_jvalue(json!([["A", "Relay1"], ["B", "Relay2"]])),
+        executed_state::par(1, 3),
+        executed_state::request_sent_by("Remote"),
+        executed_state::par(2, 0),
+        executed_state::stream_string("test", "void"),
+        executed_state::request_sent_by("Relay2"),
+    ];
 
-    assert_eq!(relay_2_res_json, relay_2_expected_json);
+    assert_eq!(relay_2_actual_trace, relay_2_expected_trace);
     assert_eq!(relay_2_res.next_peer_pks, vec![String::from("B")]);
 
     let client_2_res = call_vm!(client_2, "asd", script, relay_2_res.data, "[]");
 
-    let client_2_res_json: JValue =
+    let client_2_actual_trace: ExecutionTrace =
         serde_json::from_slice(&client_2_res.data).expect("interpreter should return valid json");
 
-    let client_2_expected_json = json!( [
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : [["A", "Relay1"], ["B", "Relay2"]]} },
-        { "call": { "executed" : [["A", "Relay1"], ["B", "Relay2"]]} },
-        { "par": [1, 3] },
-        { "call": { "request_sent_by" : "Remote" } },
-        { "par": [2, 0] },
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : "test" } },
-    ]);
+    let client_2_expected_trace = vec![
+        executed_state::stream_string("test", "void1"),
+        executed_state::stream_jvalue(json!([["A", "Relay1"], ["B", "Relay2"]]), "void2"),
+        executed_state::scalar_jvalue(json!([["A", "Relay1"], ["B", "Relay2"]])),
+        executed_state::par(1, 3),
+        executed_state::request_sent_by("Remote"),
+        executed_state::par(2, 0),
+        executed_state::stream_string("test", "void"),
+        executed_state::stream_string("test", "void3"),
+    ];
 
-    assert_eq!(client_2_res_json, client_2_expected_json);
+    assert_eq!(client_2_actual_trace, client_2_expected_trace);
     assert_eq!(client_2_res.next_peer_pks, Vec::<String>::new());
 }
 
@@ -214,7 +210,7 @@ fn join() {
                 (seq
                     (call "Remote" ("920e3ba3-cbdf-4ae3-8972-0fa2f31fffd9" "get_users") [] members)
                     (fold members m
-                        (par 
+                        (par
                             (seq
                                 (call "Relay1" ("identity" "") [] $void)
                                 (call "A" ("fgemb3" "add") [m] $void3)
@@ -233,21 +229,21 @@ fn join() {
     let relay_1_res = call_vm!(relay_1, "asd", script.clone(), remote_res.data, "[]");
     let client_1_res = call_vm!(client_1, "asd", script, relay_1_res.data, "[]");
 
-    let client_1_res_json: JValue =
+    let client_1_actual_trace: ExecutionTrace =
         serde_json::from_slice(&client_1_res.data).expect("interpreter should return valid json");
 
-    let client_1_expected_json = json!( [
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : [["A"], ["B"]]} },
-        { "par": [2, 3] },
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : "test" } },
-        { "par": [2, 0] },
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : "test" } },
-    ]);
+    let client_1_expected_trace = vec![
+        executed_state::stream_string("test", "void1"),
+        executed_state::scalar_jvalue(json!([["A"], ["B"]])),
+        executed_state::par(2, 3),
+        executed_state::stream_string("test", "void"),
+        executed_state::stream_string("test", "void3"),
+        executed_state::par(2, 0),
+        executed_state::stream_string("test", "void"),
+        executed_state::stream_string("test", "void3"),
+    ];
 
-    assert_eq!(client_1_res_json, client_1_expected_json);
+    assert_eq!(client_1_actual_trace, client_1_expected_trace);
     assert_eq!(client_1_res.next_peer_pks, Vec::<String>::new());
 }
 
@@ -313,41 +309,41 @@ fn init_peer_id() {
         ""
     );
 
-    let client_1_res_json: JValue =
+    let client_1_actual_trace: ExecutionTrace =
         serde_json::from_slice(&client_1_res.data).expect("interpreter should return valid json");
 
-    let client_1_expected_json = json!( [
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : [["A"], ["B"]]} },
-        { "par": [2, 3] },
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : "test" } },
-        { "par": [2, 0] },
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : "test" } },
-        { "call": { "request_sent_by" : "A" } },
-    ]);
+    let client_1_expected_trace = vec![
+        executed_state::scalar_string("test"),
+        executed_state::scalar_jvalue(json!([["A"], ["B"]])),
+        executed_state::par(2, 3),
+        executed_state::scalar_string("test"),
+        executed_state::scalar_string("test"),
+        executed_state::par(2, 0),
+        executed_state::scalar_string("test"),
+        executed_state::scalar_string("test"),
+        executed_state::request_sent_by("A"),
+    ];
 
-    assert_eq!(client_1_res_json, client_1_expected_json);
+    assert_eq!(client_1_actual_trace, client_1_expected_trace);
     assert_eq!(client_1_res.next_peer_pks, vec![initiator_peer_id.clone()]);
 
     let initiator_1_res = call_vm!(initiator, initiator_peer_id, script, client_1_res.data, "");
 
-    let initiator_1_res_json: JValue =
+    let initiator_1_actual_trace: ExecutionTrace =
         serde_json::from_slice(&initiator_1_res.data).expect("interpreter should return valid json");
 
-    let initiator_1_expected_json = json!( [
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : [["A"], ["B"]]} },
-        { "par": [2, 3] },
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : "test" } },
-        { "par": [2, 0] },
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : "test" } },
-        { "call": { "executed" : "test" } },
-    ]);
+    let initiator_1_expected_trace = vec![
+        executed_state::scalar_string("test"),
+        executed_state::scalar_jvalue(json!([["A"], ["B"]])),
+        executed_state::par(2, 3),
+        executed_state::scalar_string("test"),
+        executed_state::scalar_string("test"),
+        executed_state::par(2, 0),
+        executed_state::scalar_string("test"),
+        executed_state::scalar_string("test"),
+        executed_state::scalar_string("test"),
+    ];
 
-    assert_eq!(initiator_1_res_json, initiator_1_expected_json);
+    assert_eq!(initiator_1_actual_trace, initiator_1_expected_trace);
     assert_eq!(initiator_1_res.next_peer_pks, Vec::<String>::new());
 }
