+++ air/tests/test_module/integration/join_behaviour.rs
